requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiosqlite==0.20.0
redis==5.0.1
python-dotenv==1.0.1
pydantic==2.10.0
httpx==0.27.0
